Assignments
Hi
I'm Rahul Ramesh from Bengaluru, sorry for the late submission.

I'm not getting how to write programs "For Loop to Print Prime Numbers in between 1 to 1000" and "Program to print X Prime Numbers using While Loop starting from 0, and take the Input of X from the User".

Please help me out. I want to learn and understand but I'm not able to analyse. It would be a great help if you get back to me.